requirements.txt
# place the requirements for the tool to function
superagi-tools==1.0.6
pydantic==1.10.8
pytest==7.4.0
pytest-mock==3.11.1